requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-taggit==3.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tweepy==4.13.0
urllib3==1.26.15
djangorestframework~=3.14.0
